Refactor NotificationManager to use useNotificationManager hook and remove unused useNotification hook
